# interfaces/http/openapi/1.0.0.json
{
  "components": {
    "schemas": {
      "FeatureRequest": {
        "description": "Payload describing the series to transform into features.",
        "properties": {
          "bars": {
            "description": "Ordered price bars.",
            "items": {
              "$ref": "#/components/schemas/MarketBar"
            },
            "minItems": 1,
            "title": "Bars",
            "type": "array"
          },
          "symbol": {
            "description": "Instrument identifier.",
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "bars"
        ],
        "title": "FeatureRequest",
        "type": "object"
      },
      "FeatureResponse": {
        "description": "Response containing the most recent feature snapshot.",
        "properties": {
          "features": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Features",
            "type": "object"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "features"
        ],
        "title": "FeatureResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarketBar": {
        "description": "Representation of a single OHLCV bar for online inference.",
        "properties": {
          "askVolume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Ask-side queue volume for microstructure features.",
            "title": "Askvolume"
          },
          "bidVolume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Bid-side queue volume for microstructure features.",
            "title": "Bidvolume"
          },
          "close": {
            "description": "Close price for the interval.",
            "title": "Close",
            "type": "number"
          },
          "high": {
            "description": "High price for the interval.",
            "title": "High",
            "type": "number"
          },
          "low": {
            "description": "Low price for the interval.",
            "title": "Low",
            "type": "number"
          },
          "open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opening price for the interval.",
            "title": "Open"
          },
          "signedVolume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Signed volume (buy-sell imbalance).",
            "title": "Signedvolume"
          },
          "timestamp": {
            "description": "Timestamp of the bar in ISO 8601 format.",
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Traded volume for the bar. Optional for illiquid venues.",
            "title": "Volume"
          }
        },
        "required": [
          "timestamp",
          "high",
          "low",
          "close"
        ],
        "title": "MarketBar",
        "type": "object"
      },
      "PredictionRequest": {
        "description": "Prediction request payload, optionally specifying a forecast horizon.",
        "properties": {
          "bars": {
            "description": "Ordered price bars.",
            "items": {
              "$ref": "#/components/schemas/MarketBar"
            },
            "minItems": 1,
            "title": "Bars",
            "type": "array"
          },
          "horizon_seconds": {
            "default": 300,
            "description": "Prediction horizon in seconds for contextual metadata.",
            "maximum": 3600.0,
            "minimum": 60.0,
            "title": "Horizon Seconds",
            "type": "integer"
          },
          "symbol": {
            "description": "Instrument identifier.",
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "bars"
        ],
        "title": "PredictionRequest",
        "type": "object"
      },
      "PredictionResponse": {
        "description": "Response representing the generated trading signal.",
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "horizon_seconds": {
            "title": "Horizon Seconds",
            "type": "integer"
          },
          "score": {
            "description": "Composite alpha score driving the action.",
            "title": "Score",
            "type": "number"
          },
          "signal": {
            "additionalProperties": true,
            "title": "Signal",
            "type": "object"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "horizon_seconds",
          "score",
          "signal"
        ],
        "title": "PredictionResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2Bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "TradePulse Platform Team",
      "url": "https://github.com/neuron7x/TradePulse"
    },
    "description": "Production-ready endpoints for computing feature vectors and generating lightweight trading signals from streaming market data.",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "TradePulse Online Inference API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/features": {
      "post": {
        "operationId": "compute_features_features_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2Bearer": []
          }
        ],
        "summary": "Generate the latest engineered feature vector",
        "tags": [
          "features"
        ]
      }
    },
...
    "/predictions": {
      "post": {
        "operationId": "generate_prediction_predictions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2Bearer": []
          }
        ],
        "summary": "Produce a trading signal for the latest bar",
        "tags": [
          "predictions"
        ]
      }
    }
  },
...
}